Handled exception

+++ main.py
@@ -18,23 +18,30 @@ def result():
    vid_path = request.form.get('vid_path')
    output_path = request.form.get('output_path')
 
-   if request.method == "POST":
-      vidcap = cv2.VideoCapture(vid_path)
-      success,image = vidcap.read()
-      count = 0
-      while success:
-         cv2.imwrite(output_path + "/frame%d.jpg" % count, image)     # save frame as JPEG file      
+   if os.path.isfile(vid_path):
+      if request.method == "POST":
+         vidcap = cv2.VideoCapture(vid_path)
          success,image = vidcap.read()
-         count += 1
-      ls = os.listdir(output_path)
-      rand = random.choice(ls)
-      result = output_path + '/' + rand
-      src = result
-      shutil.copy(src, app.static_folder)
-      return render_template('result.html', result = rand)
+         count = 0
+         while success:
+            cv2.imwrite(output_path + "/frame%d.jpg" % count, image)     # save frame as JPEG file      
+            success,image = vidcap.read()
+            count += 1
+         ls = os.listdir(output_path)
+         rand = random.choice(ls)
+         result = output_path + '/' + rand
+         src = result
+         shutil.copy(src, app.static_folder)
+         status = 'Frames are saved to the given path!'
+         return render_template('result.html', result = rand, status=status)
+      else:
+         result = ''
+         status = 'Some thing went wrong!!'
+         return render_template('result.html', result=result, status=status)
    else:
-      result = 'Some thing went wrong!!'
-      return render_template('result.html', result=result)
+      result = ''
+      status = 'Something went wrong!'
+      return render_template('result.html', result=result, status=status)
 
 if __name__ == '__main__':
    app.run(debug = True)
